chore: node components moved to separate folder
