+++ frontend/src/components/node/trigger-node.tsx
